move view_user_vaults tests to test_vaults.py
